refactor: make rate limiting a reusable helper func
